chatback/pyproject.toml
[tool.poetry]
name = "chatback"
version = "0.1.0"
description = "Chat backend for AI4MDE"
authors = ["Mimo Amghar <mimo@amghar.io>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.115.6"
uvicorn = "^0.32.1"
sqlalchemy = "^2.0.36"
pydantic = "^2.10.3"
pydantic-settings = "^2.6.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["argon2"], version = "^1.7.4"}
argon2-cffi = "^23.1.0"
python-multipart = "^0.0.19"
psycopg2-binary = "^2.9.10"
asyncpg = "^0.30.0"
langchain = "^0.3.11"
langchain-openai = "^0.2.12"
langchain-community = "^0.3.11"
redis = "^5.2.1"
qdrant-client = "^1.12.1"
email-validator = "^2.2.0"
alembic = "^1.14.0"
pytz = "^2024.2"
langchain-core = "^0.3.24"
langchain-redis = "^0.1.1"
httpx = "^0.28.1"
tenacity = "^9.0.0"
termcolor = "^2.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py312']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi-line-output = 3